+++ app/diary/_components/DiaryDetail/ReComment/index.tsx
@@ -3,7 +3,7 @@ import * as styles from "./style.css";
 import Image from "next/image";
 import LikeIcon from "@/public/icons/like.svg";
 
-const ReComment = () => {
+const ReComment = ({ reply }) => {
   return (
     <div className={styles.replyContainer}>
       <div className={styles.commentHeader}>
